+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Loop tag names trimming.
 - Custom loop delimiters support.
+- Zip export and typings.
 
 ## [0.7.3 - 2019-10-11](https://github.com/alonrbar/easy-template-x/tree/v0.7.3)